pyproject.toml
[project]
name = "mt-gbm"
version = "0.1.0"
description = "Multi-Task GBDT Implementation for CVR Prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.2.0",
    "tensorflow>=2.12.0",  # メモリ制御等の最適化機能で使用
    "pyyaml>=6.0",  # ハイパーパラメータ設定のYAML読み込み用
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]